drop providers for model info and rely on native columns data
